Only set Location header on Post files not content-location
Since Content-Location refers to the returned representation in this context, which can't be directly fetched from any endpoint (has to be a part of fetching the message object), this shouldn't be included.
On the other hand, Location refers to the primary resource being created, which in this context is deemed to be the file that's been uploaded, therefore this can and should still be included.